commit to remote

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: jylimaul <jylimaul@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/28 13:00:31 by jylimaul          #+#    #+#             */
-/*   Updated: 2021/11/17 13:36:49 by jylimaul         ###   ########.fr       */
+/*   Updated: 2021/11/19 11:55:29 by jylimaul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len] != '\0')
